+++ src/ts/infrastructure/ui/index.ts
@@ -10,3 +10,5 @@ export * from './loadingButtonState';
 export * from './CustomSnackbar';
 export * from './dbButton';
 export * from './reactRoot';
+export * from './tabController';
+export * from './navDrawer';